requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]
pyjwt
pytest
httpx
pytest-asyncio
